services/dq-gate/requirements.txt
kafka-python==2.0.2
prometheus-client==0.20.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
jsonschema==4.23.0
PyYAML==6.0.2
opentelemetry-semantic-conventions==0.48b0
